Change the type that holds the value in Myingameport to int, so that it doesn't lose information.
git-svn-id: https://svn.code.sf.net/p/fmslogo/source@65 9d0050d7-00b0-40f9-871c-8e02cbd5794f

+++ trunk/logo/src/assembly.cpp
@@ -120,7 +120,7 @@ short Myinport(short portid)
 
 int Myingameport(short portid, short mask)
    {
-   short value = -1;
+   int value = -1;
 
    JOYINFOEX joystickInfo;
    joystickInfo.dwSize  = sizeof joystickInfo;
